requirement.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]

##-e .